removed InstanceRegistry
this resolves ALF-2, since deleted code can't be untested 🙃.

instead of using the instanceRegistry, the _programs mapping is used.
Since templates can't be at address(0), the value of the template field
is used to determine if the program is registered or not.

A lot of methods (unnecessary for on-chain things AFAIK) were removed,
so this probably breaks an existing frontend if there is such a thing