implement feedback from budi
- removed commented & redundant test
- fixed test_update_nonexistent

the EnumerableSet is a library, and it's inlined to the contract itself
(I can be sure of it because the methods of the `EnumerableSet` are
`internal`).

the try/catch construct only works with external calls, so using it
as-is isn't an option.

The expectRevert I guess works in the same way. I mean the docs say it
inverts the behaviour of the revert check on the next call, so there's
no way it can work with an internal call

also, we were trying to match against a bytes4 object in the
expectRevert, but the error itself was longer since it had the 0x11
parameter

To fix this, I added a wrapper contract around the set in order to force
an external call.