remove enumerableset
turns out it was pretty easy, and the behaviour of returning the size of
the template set wasn't explicitly tested so it won't be missed.

having the template field in the struct is redundant, since it's also
used as the key.

However, strings always take up 1 word of storage, and the other slot is
shared by the address (20 byte) and bool (1 byte), so using a smaller
datatype such as an 'exists' boolean, doesn't actually save up space.

If the compiler was smart enough to store strings differently when it's
useful to squeeze an extra bit or two in it, it'd be a different story.