tests for errors in initialization
the tests were formerly skipped, and now are re-enabled

the main issue was that in AludelFactory we create the contracts with
the ProxyFactory, which calls the initialization method if params for it
are provided.

The ProxyFactory then checks for success in the initialization call, and
it calls `require(success, string(err))`. This casts the custom error of
the Aludel into a string, the representation of which seems to break
hardhat-chai-matchers way of parsing a transaction's return values.
Sorry, but I missed that in the audit 🙃.

The workaround for this is to both do the call and catch the error
manually, and assert against the cursed return value.