add user manager and user contracts
the goal of these contracts is abstract the user interaction in the tests.
The user instance contains its private key which is used to sign messages, such as Lock and Unluck permissions.
The user manager derive a key and create new user instance with it.

remove PRIVATE_KEY constant

rename UserManager to UserFactory