manage all dependencies via npm
we were having some compilation issues in foundry due to using an old
version of forge-std. The version in ./lib was up to date, but the one
in node_modules wasn't. I was very convinced foundry was using the files
in ./lib but that wasn't the case. Wasted like 1hr of my time.