added a semver-range for big-name dependencies
rationale for this is that well established dependencies can usually be
trusted to respect semver versioning and not push breaking changes on a
patch release. smaller ones, or ones for which we're using a 0.something
or beta release, should be fixed to a particular version.

hardhat itself, however, I had to pin to keep the project working.
Later I'll remove viaIR to get a more 'normal' hardhat